requirements.txt
django==3.2.10
pymysql==1.0.2
djangorestframework==3.13.1
coreapi==2.3.3
djangorestframework-jwt==1.11.0
django-cors-headers==3.10.1
pika==1.2.0
oss2==2.15.0
django-filter==21.1